requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.2
coloredlogs==15.0.1
dataclasses-json==0.5.8
dnspython==2.3.0
duckdb==0.8.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
flatbuffers==23.5.26
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
humanfriendly==10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.197
langchainplus-sdk==0.0.8
lz4==4.3.2
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
onnxruntime==1.15.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
orjson==3.9.1
overrides==7.3.1
packaging==23.1
pandas==2.0.2
posthog==3.0.1
protobuf==4.23.2
pulsar-client==3.2.0
pydantic==1.10.9
pypdf==3.9.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
zstandard==0.21.0
